refactor the dummy core
