+++ test_interconnect/test_circuit.py
@@ -3,33 +3,7 @@ from interconnect.cyclone import PortNode, SwitchBoxNode, SwitchBoxSide
 from interconnect.cyclone import SwitchBoxIO
 from interconnect.sb import DisjointSB, WiltonSB, ImranSB
 from interconnect.tile_circuit import TileCircuit, SBConnectionType
-from common.core import Core
-import magma
-
-
-class DummyCore(Core):
-    def __init__(self):
-        super().__init__()
-
-        # PEP 8 rename
-        t_data = magma.Bits(16)
-
-        self.add_ports(
-            data_in=magma.In(t_data),
-            data_out=magma.Out(t_data),
-        )
-
-        self.wire(self.ports.data_in, self.ports.data_out)
-
-    def inputs(self):
-        return [self.ports.data_in]
-
-    def outputs(self):
-        return [self.ports.data_out]
-
-    def name(self):
-        return "DummyCore"
-
+from .dummy import DummyCore
 
 def test_connection():
     """test low-level add/remove connections"""
