+++ test_interconnect/test_interconnect.py
@@ -5,32 +5,7 @@ from interconnect.cyclone import SwitchBoxSide, SwitchBoxIO
 from interconnect.util import create_uniform_interconnect
 from interconnect.interconnect import TileCircuit
 from interconnect.sb import SwitchBoxType, DisjointSB
-from common.core import Core
-import magma
-
-
-class DummyCore(Core):
-    def __init__(self):
-        super().__init__()
-
-        # PEP 8 rename
-        t_data = magma.Bits(16)
-
-        self.add_ports(
-            data_in=magma.In(t_data),
-            data_out=magma.Out(t_data),
-        )
-
-        self.wire(self.ports.data_in, self.ports.data_out)
-
-    def inputs(self):
-        return [self.ports.data_in]
-
-    def outputs(self):
-        return [self.ports.data_out]
-
-    def name(self):
-        return "DummyCore"
+from .dummy import DummyCore
 
 
 def test_tiling():
